backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.5.4
google-api-core==1.34.1
protobuf==3.20.3
grpcio==1.48.2
grpcio-status==1.49.0rc1
mysql-connector-python==8.0.34